JS improvements
* List the .qml / .js files found in the library as process presets
* Change the JS API (sorry).

The tick function now matches the C++ graph_node API.
See the classes
- TokenRequestValueType
- ExecutionStateValueType

to see the available properties. In particular notice that we now have access
to musical metrics information from the JS scripts, as well as to the sample rate, etc.

It is recommended to read ossia::token_request documentation to see all the nice features available,
e.g. find when is the next quantification date for a given musical duration, etc...